requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
mediapipe>=0.10.13
pandas>=2.1.1
numpy>=1.24.3
scikit-learn>=1.3.2
python-dotenv==1.0.0
